requirements.txt
google-adk>=0.1.0
google-generativeai>=0.3.0
litellm>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
python-dateutil>=2.8.2
tzdata>=2023.3
backports.zoneinfo>=0.2.1;python_version<"3.9"
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0 